一部関数をcmd_cmd_invocation2.cに移動
